feat: alertdialog tests
Add HTML patterns for an 'alert' dialog.

Does the `role="alertdialog"` attributes/value works as expected for a cookie message dialog that is present when the page loads?

We want to see if the user is told about the alert and is guided towards it whilst other page elements are obscured