feat: Loading indicator experiments
Pages show loading indicators to let a user know _something_ is happening in the background and that they should wait for it to finish. This works well for users who can see the visual indicator but what about those who cannot: we need to let screen-readers know something is happening too.